Make RaftConsensus::Update() snooze the timer when waiting for its own log appends
Right now elections are extremely churny because log appends can take a long time
and while we're waiting for those to complete we don't reset the timer, making
leader elections fire.

This just changes this so that we keep snoozing the timer until the log append completes.
When we get rid of 'update_lock_' (KUDU-567) this should no longer be necessary.

Change-Id: I851fbe82ce69356280f0d99d2275fb04110a58a3
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5575
Tested-by: Todd Lipcon <todd@cloudera.com>
Reviewed-by: Todd Lipcon <todd@cloudera.com>