[java client] Make KuduTableInputFormat leader election-resistant
As seen when running the ITBLL against a cluster that was getting nodes randomly killed,
it's possible to get transient failures when starting a job that scans Kudu since it
assumes being able to get a leader per tablet. This patch will make it so that we retry
getting the list until we're given a leader per tablet, else we'll pick the first one and
let the client deal with it (puts sunglasses on).

Change-Id: Iaf25c0125c3d19ea0b426271e3fdb5e2334fe9e8
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5581
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>