MarkReplicaFailed() should return bool
Instead of doing a DCHECK in here when the specified TS doesn't actually
host a given tablet, we should simply return true/false depending on
whether we found the tablet on that TS or not. This fixes a regression
where MetaCache::MarkTSFailed() would crash due to trying to mark the TS
failed for every tablet (admittedly a simple approach). Returning bool
still allows us to DCHECK in certain places where we want to ensure that
we are using a consistent view of the metadata.

Also, make logging in client-test a little more verbose so it's easier
to follow along with.

Change-Id: I205c4d5a477b2479ed1c99b3207eb23fe22c21de
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5585
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: jenkins