Rename client shell-script tests to end in .sh
This seems to be a little more obvious since this is the only place I
know of that we have shell-script unit tests. I thought somehow the
build was putting binaries in my source dirs.

Also, when the tests are run, use their name not including their file
extension for the log output file.

Change-Id: I1e79fbf6bf834aaeed38dad95f780ce89af7ae68
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5586
Tested-by: jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>