thirdparty: better handling for msan_interface.h in llvm
Between LLVM 3.4 and 3.5, the handling of msan_interface.h changed somewhat.
See commits eca6314 (projects/compiler-rt) and b584e19 (llvm proper) for
more details. Because of these changes, LLVM 3.5 cannot be used to build our
LLVM 3.4.2 in thirdparty. A bunch of LLVM 3.4 binaries fail to link with
errors "error: undefined reference to '__msan_allocated_memory'".

To fix this, let's patch LLVM 3.4.2 with commit b584e19. This will remove
the HAVE_SANITIZER_MSAN_INTERFACE_H check completely (obviating the need for
our workaround in codegen) and replace it with a
__has_feature(memory_sanitizer) check instead, which is only true when
building with LLVM and when -fsanitize=memory is set.

Change-Id: I9e7d7496a40adc28607d98bff8e32d8f8ab1bca4
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5592
Reviewed-by: Vladimir Feinberg <vladimir.feinberg@cloudera.com>
Tested-by: jenkins