Increase log retention in the test clusters
Per David:
We don't have enough log retention. we kill a node and when it is finally done bootstrapping the
current leader might not have the log entries it needs to get up to speed. A leader will hang on
to entries for killed node, but if the leader changes then all bets are off. There's a flag for
this: "log_min_segments_to_retain" we need to turn this dial up. The default is 2, in this case
we would need something like 3, I suggest you turn the dial to something like 10 (if we can
spare the space).

Change-Id: I280bedac2fe383ed55dc548dc30df28a16742895
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5599
Tested-by: David Alves <david.alves@cloudera.com>
Reviewed-by: David Alves <david.alves@cloudera.com>