Removed unused single-row-add methods to RowBlock protobuf conversion.
The single-row addition methods allowed one to add rows to the RowBlock
protobufs one row at a time, but because of this they were inefficient.

Their use was discouraged in the documentation and no one actually called
the functions outside of the wire_protocol's one unit test. For this reason,
they were removed.

To increase testing coverage the RowBlock conversion function's test size
was increased.

Change-Id: I312fb3cfd4828c2f01163bf1525827b0ddb4bb6a
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5609
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>