Make RaftConsensusItest run faster in non-slow mode
This just reduces the number of writes/elections/replicas in non-slow
mode.

Without this patch, in non-slow mode, the suite would take 70-100 secs
now it clocks in the range of 24-28 secs most of the time. The "offending"
tests (the failovers and the automated election test) clocked at about
40 and 8 secs before. They they clock in the ranges of 5-8 and 5-6 secs,
respectively, now.

As discussed in gerrit, even though these times are still a bit high,
particularly for the failovers test, these are end-to-end failover tests
that cover many components and should be part of tests we run locally.

Change-Id: I39c5c1e5f05414570dbdc494530c450da6faf177
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5613
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: David Alves <david.alves@cloudera.com>