KUDU-583. Flush MemRowSets before scanning in full stack test.
By flushing the MemRowScans before performing the scan benchmarks in the
full stack test, we are able to more accurately model the actual performance
that will be seen in a real-life application. Most of the time the data is
stored on disk, rather than in memory.

The issue was that the MaintananceManager could very well have been behind
in the flushing operation at the time of running the scans, so the
benchmark ended up recording the times with in-memory scanning.

Perf differences:
http://github.mtv.cloudera.com/gist/vlad17/e0daab71af4eb9938ff3

Change-Id: Ia5d11272a1f06f8cd086628da076630e3c05e93a
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5616
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>