Make failure handling, on transaction abort, run on the apply pool
We're handling transaction aborts in the caller thread while we handle
the applies in their own threads. This just makes aborts take the same
thread as applies.

The reason for this change is that we had a possible lock inversion race
that could result in a deadlock, where on transaction abort consensus
would acquire the consensus lock and then indirectly the tablet manager
lock, and when reporting the reporting thread would acquire the same
locks in reverse order.

Change-Id: I7611a2a61795efb172b9d9dd221e511f42ddd032
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5618
Tested-by: jenkins
Reviewed-by: Michael Percy <mpercy@cloudera.com>