jenkins: Fix handling of unparseable error messages
The parse_test_failure.py script currently doesn't know how to parse
generic c++ assertion failures. We'll add that feature in a future
commit. For now, ensure that we do a better job of generating valid XML
when we don't know what caused a failure and also of providing some kind
of explanation when that happens.

Now, when we see a "start unit test" log line that is not matched by an
"end unit test" log line, and no other errors were parsed from the log
for that unit test, we print the following error message:

Unrecognized error type. Please see the error log for more information.

Change-Id: Ie53fcc062d33291d05e7a86505efb9f062b56440
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5660
Tested-by: Michael Percy <mpercy@cloudera.com>
Reviewed-by: Todd Lipcon <todd@cloudera.com>