Bug Fix: Enforce ordering on bootstrap
We currently have a bug where we would apply replicates out of order even if they
touch the same rows.

Normally serialization ordering is enforced by preparing, but on bootstrap we prepare
and apply in one shot. This, coupled to the fact that we apply commits as we see them,
means that we may apply two transactions that touch the same rows out of order.

This became apparent when mutations along with flushes/compactions were enabled in
linked_list-test. The case we were seeing was the following:
- Replicate 1.3340 (..., Insert row *67, ...)
- ...
- Replicate 1.3342 (..., Mutate row *67, ...)
- Commit 1.3342
- Commit 1.3340

Applying Commit 1.3342 would fail with an error indicating the mutated row did not exist.
The solution that this patch implements is to buffer commits and only apply them in order.
An alternative approach would be to separate prepare and apply, on replay, but this would
require threading and would be more complex.

This also includes a test that would fail before but passes now.

Change-Id: I6ee2315083d9e410a1936c1dbc6b27dba0f2a03a
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5671
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>