Stop pre-allocating MRS and DMS's arenas
Right now we pre-allocate those arenas to 1.5MB and 1MB, which means that when we check their
memory footprint it shows at least that amount of data. It's mostly fine with MRS but we have
one DMS per row set so it can add up enough that it's over the flush size, without any data.

It also means we pre-allocate all those 1MB for all the DMS, which can amount to hundreds of
MBs with enough tablets and row sets (as seen on the YCSB cluster).

This patch is pre-allocating 16 bytes instead of 0 because of this error and it's NYE:
F1231 15:59:51.898149 18789 arena.cc:108] Check failed: reinterpret_cast<uintptr_t>(buffer->data())
& (16 - 1) == 0 (11 vs. 0) Components should be 16-byte aligned: 0x7fe4e79eec2b

Change-Id: I676f8d7641c9286dd5d03b7dbc3a014a748d6c8b
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5701
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>