rpc: fix OutboundTransfer to properly account for bigger sidecars
This fixes a bug in how OutboundTransfer was creating iovecs when dealing with
larger responses, which would typically get split across several sendmsg()
syscalls. The bug would cause the server to think it had completed sending the
response when in fact it had only send a prefix. This caused the client side to
get stuck waiting for the rest of the response to appear.

The newly added test would fail with a timeout before the bug fix, and now
passes. This should fix an issue seen with the new RPC sidecar support on
scans when the scan response is large (eg the tpch1 benchmark job).

Change-Id: I9cff60356518a07ca857f37f7d4463e511cfe3c8
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5714
Tested-by: Todd Lipcon <todd@cloudera.com>
Reviewed-by: Vladimir Feinberg <vladimir.feinberg@cloudera.com>