KUDU-454. rpc: fix flaky TestCallTimeout
This fixes a rare case where a call could time out during negotiation, but
result in a NetworkError instead of a TimedOut status. This is hard to
reproduce, but looking at the logs from one of the failed test runs, I can see
that this is the case that was triggered.

Change-Id: I9cb716b5f1ddc24246253d9e2ce3325925dd98bd
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5715
Tested-by: jenkins
Reviewed-by: Michael Percy <mpercy@cloudera.com>