Reuse the same row operation error status, flushed status, across replays and update the stats properly
On the first replay after a crash, when we find a flushed insert/mutation, we set the row
operation as failed with Status::AlreadyPresent() and update the "ignored_*" counters.
However, when replaying for a second time we change that error to be a Status::RuntimeError(),
making the error differ across replays unecessarily. Moreover we do not update the ignored
counters properly on the second replay.

This patch fixes this so that we reuse the error on replay, but adds a "flushed" bool field
to be used instead of Status::AlreadyPresent() to signal that an operation was flushed in a
previous replay. This also updates the "ignored_*" counters when we see an operation which
has been previously flushed.

Change-Id: Ie7a4e119a45010ab9e36cbd8cb288e60cc75d85a
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5723
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>