KUDU-402: ease TSAN deadlocks in subprocess child
KUDU-402 shows that TSAN-instrumented calls after fork() but before exec()
can cause the child to deadlock. This commit moves the call to
prctl() so that it's the very first call following the fork(). As long as it
doesn't trigger a deadlock in and of itself, this should guarantee that when
the parent (the test) times out, the child is killed. Meaning, this should
convert outright build timeouts into test timeouts.

I've also switched the signal SIGINT to SIGTERM, which is more appropriate
for abrupt process termination.

Change-Id: Ieafcc7456a703501653ae801b563148de938aa6d
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5727
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Tested-by: jenkins