KUDU-402: use the latest clang in the PATH for ASAN/TSAN/CLIENT builds
Since we can't upgrade thirdparty/llvm without also building in C++11 mode,
let's take a different tack: decouple the clang used to build Kudu from the
llvm found in thirdparty. The headers and libraries of the latter will
continue to be used for codegen while the former will be used to actually
compile the Kudu codebase.

Change-Id: I9e15303108c38c5d61dc9434f1efd1dc0c90c7cf
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5738
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Tested-by: jenkins