log: fix GCing of log index chunks based on segment retention
This fixes a bug we saw in the integration cluster where we would use the
minimum needed OpId to GC the log index chunks regardless of how many segments
should be retained.

For example, we might be configured to retain 10 segments worth of logs (so
that other peers can catch up), but we may have fully flushed all of our
in-memory data. In that case, the min_op_idx passed to Log::GC() may fall in
the latest log index chunk, and cause us to remove the earlier one.

Change-Id: I9c810b04bc958b539d20c489d4831e07431f942c
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5775
Reviewed-by: Jean-Daniel Cryans <jdcryans@cloudera.com>
Tested-by: Todd Lipcon <todd@cloudera.com>
Reviewed-by: Michael Percy <mpercy@cloudera.com>