Replace the MM's anchored ms concept with log retention size
Currently we have a ts_anchored_ms field in the MM ops that represents the age of the oldest
transaction that the op anchors. Time is a tricky thing, and it's best to use data size as
it's more actionable. For example, two different ops could both anchor transactions that are
5 minutes old, but one could be preventing us to GC tens of MBs whereas the other one could
just be a few MBs (let's say because it's a less write-heavy tablet).

This patch thus replaces ts_anchored_ms with logs_retained_mb.

It also gets rid of the related max config in the MM, it wasn't used and a replacement isn't
really needed (could revisit later).

Finally, this patch improves how we select an MM op. After we decide that we don't need to flush
based on anchored RAM, we check if any task is preventing us from GCing logs. The one that
prevents the GCing of the most MBs win, unless another task has an equal amount of
logs_retained_mb, but also has more RAM anchored.

Change-Id: I47c1fe962d72326ceb2680a68acb1b9c23a69be3
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5780
Tested-by: Jean-Daniel Cryans <jdcryans@cloudera.com>
Reviewed-by: Michael Percy <mpercy@cloudera.com>