Add trace-viewer to thirdparty
This adds a built version of the trace-viewer HTML and JS to thirdparty
via our S3 mirror. I included instructions on how to build it, and the
build script itself provides some instructions how to upload to S3.

The particular branch of trace-viewer being built is slightly customized
to Kudu -- I removed some Chrome-specific stuff and cloned it into its
own part of the trace-viewer repository tree. If we see fit, we can extend
it with Kudu-specific object visualizations, etc.

Tested the result by starting a TS and loading /tracing.html

Change-Id: I0eaaaca925ec364da03d9fa15df0834935a23735
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5783
Tested-by: jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>