KUDU-452: eagerly open data blocks during remote bootstrap
Opening a block anchors it; only when the last reader is closed is a
block deletion actually effected. Thus, all remote bootstrap needs to do in
order to safely operate on data blocks is to eagerly open them all when a
RB session is begun.

One side effect of this change is that we can no longer eagerly close data
blocks because we can't guarantee that they can be reopened (i.e. if they
were since deleted).

Change-Id: I277f911166145b8f242c793d471b7bdcbb18399f
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5793
Tested-by: Adar Dembo <adar@cloudera.com>
Reviewed-by: Michael Percy <mpercy@cloudera.com>