KUDU-25: pbc-dump should, um, actually work
Apparently I had been testing with old build output and didn't notice when I
totally broke pbc-dump.

The problem: a single schema isn't enough to fully describe a message; we
need to include the message's dependent schemas too. As before, using protoc
to get at the entire FileDescriptorSet is lame when we have all the
information we need at run time. But it's more complicated: we need to walk
a message's dependency graph and include all schemas along the way in the
container's supplemental header.

If this walk proves too expensive, it's trivial to cache it; the PB schema
information is static and won't change after the server has been started.

To ward against future breakages, I've refactored pbc-dump so that the
interesting functionality can be unit tested.

Change-Id: Ia33aab915ac8ba40cda0424f0f182a8a521fc166
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5808
Tested-by: jenkins
Reviewed-by: Michael Percy <mpercy@cloudera.com>