Style cleanup of some wire protocol helpers
We adding in-params at the end of the function signature when adding
sidecars to the RPC impl. Putting the out-params back at the end per the
style guide.

Also renamed the helper function RewriteRowBlockPB() to
RewriteRowBlockPointers() to be more accurate and added a few comments
in the header and impl.

Change-Id: I6d822d19d8637a345191388f077b5d71e3c001de
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5889
Reviewed-by: Andrew Wang <andrew.wang@cloudera.com>
Tested-by: jenkins