KUDU-601: Fix RemoteKscTest.TestMasterOk in TSAN
GetLeaderMasterRpc::SendRpc() was releasing 'lock_' before
all of the sub-RPCs were sent, allowing a condition where:

1) Thread 1 sends first sub-RPC and releases 'lock_'.
2) Thread 2 receives the response from the first sub-RPC, runs
the callback, discovers the new leader, and deletes the RPC.
3) Thread 1 tries to evaluate the loop condition and triggers
use-after-free (because the condition includes RPC state that
was deleted in step 2 above).

Sending all of the sub-RPCs while 'lock_' is held fixes the
issue.

Change-Id: I9425887dd9975805c43c65037a93e79e380fbf55
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5869
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: jenkins