remote bootstrap: always use new block IDs for copied blocks
The existing RB client tried to reuse the downloaded block's ID when
creating a local block. Apart from not being collision-proof (what if I
bootstrap a tablet, delete its metadata, and bootstrap it again?), this
hampers the ability of the block manager to use arbitrary block ID schemes.
For example, if a block ID includes a unique disk identifier, it will always
be unportable across machines.

So instead, let's make the RB client always use new block IDs. The only
gotcha is that we have to rewrite the superblock to replace the old IDs with
the new ones.

The new unit test I added uncovered a bug in RemoteBootstrapSession.

Change-Id: I233ad6d9a70a62deecab0145d25f941c51ce83f4
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5876
Tested-by: jenkins
Reviewed-by: Michael Percy <mpercy@cloudera.com>