tablet: do not delete old superblock when flushing new one
There's just no need:
1. Flushing the new sb will atomically replace any on-disk sb file.
2. When a tablet is loaded, any stale sb is ignored.
3. Omitting the deletion avoids an unnecessary directory fsync.

Not to mention that asking the block manager (via the fs manager) to
delete the sb is wrong since the block manager never created it in the first
place. It happens to work when using the file block manager since the
legacy sb creation code in fs manager uses the same file layout, but with
the log block manager, the sb will never be found.

Change-Id: Ibcff65cb2419633d6ad8167bfc8986003ecacca8
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5895
Tested-by: jenkins
Reviewed-by: Michael Percy <mpercy@cloudera.com>