Set leader Master host/port from specified addresses.
Sets the leader Master host/port based on the original addresses
passed to GetLeaderMasterRpc, as opposed to from the RPC address or
hostname retrieved from GetMasterRegistration. This is already done
by the Java client and allows for handling of situations where
a Masters are bound on multiple addresses, and we want the TS and clients
to connect only to specific addresses (e.g., if the Master have multiple
NICs and we want the clients to connect over one specific NIC).

Change-Id: I540736092b55f1e5222270c1393c430291764203
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5899
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: jenkins