Add a way to identify log segments that we can't GC
An issue we currently have is that we have an indirect control over the number of
log segments. We make sure not to go down below a certain level, we stop GCing at a point,
but we have no way to cap it.

This patch adds a way to get the layout of the log segments that we can't GC so that we later
use it to calculate log retention for MRS and DMS.

Change-Id: Ica4b0876ca4da824df80f5fc9014df1a0071bed4
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5729
Tested-by: jenkins
Reviewed-by: Michael Percy <mpercy@cloudera.com>