Pass the log retention information to the Maintenance Manager
Now that the MM understands how to prioritize log retention, and that we're able to tell how much a
rowset is retaining, we put the two together.

The only tricky thing in this patch is that we don't have a MM operation per DMS, there's too
much churn. Instead, we just report the stats for the best candidate to flush, then if the MM
selects a specific tablet then it will re-evaluate which DMS should really be flushed.

Change-Id: I2def2ab61fe630dea9b5fbca69c8db8602a29075
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5875
Reviewed-by: Michael Percy <mpercy@cloudera.com>
Tested-by: jenkins