fs: basic block manager multi-path support
This commit implements basic multi-path support in both the log and file
block managers. In both cases we're using a simple round-robin strategy
to place the next block or container; in the future block placement hints
will help with locality.

The changes to the log block manager were quite simple as there's already a
large map of all blocks to their respective locations. Open() looks for
existing containers across all root paths, and CreateBlock() cycles through
the root paths when it needs to create a new container.

The file block manager implementation is more complicated because it lacks
such a map, and I think having one would go against its design principles
(i.e. rely on the filesystem for as much metadata management as possible).
Instead, we take advantage of the fact that block identifers are opaque and
build the path information directly into each block's ID. First, we generate
and persist a UUID for each path; this also helps us deal with disks that
get remounted. Next, we maintain a map of UUIDs to paths for fast lookups.
Lastly, we prepend the path UUID to each block ID in CreateBlock(). The
FileBlockLocation class encapsulates this new block ID complexity.

Of note:
- Due to the legacy metadata block support in FsManager, BlockId is still
  rather messy. It will be cleaned up eventually.
- block_manager.h now emphasizes that Create() should be followed by Open().
  The unit tests also reflect that.
- block_manager-stress-test can now be run with multiple paths.
- Having realized that we may have millions of FileReadableBlock instances,
  I removed the 'closed_' field to save some space. It's only one byte,
  though that balloons to eight bytes with padding.
- FsManager still exposes one path, which means the rest of Kudu isn't able
  to use multiple paths just yet. That'll come next.
- The log block manager writes metadata instance files too. Seems like a
  nice thing for debugging.

Change-Id: I4b4185181064f50b5354b9695d5386259a8e4838
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5958
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>