fs: do not fsync parent directories when deleting blocks
Mike and I could not fathom a case where making a deletion durable
immediately is required vs. opposed to letting the filesystem make
everything durable every 5 seconds or so. So let's not do that.
Granted, the "delete garbage blocks on startup" logic mentioned in the
comment doesn't exist yet, but we know we need it.

The log block manager is already in good shape, as deleting a block means:
1. Write a record describing the deletion to the container's metadata file.
2. Sync the metadata file.
3. When the last reader closes, punch a hole in the container's data file.

The fsync() in step #2 is required, otherwise the hole punch may be made
durable before the metadata record is made durable.

Change-Id: I146123ace8d6b0f4980e2568a506a866282b2b3b
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5936
Reviewed-by: Michael Percy <mpercy@cloudera.com>
Tested-by: Michael Percy <mpercy@cloudera.com>