fs: expose wal dir and data dirs to the rest of Kudu
Building on the multi-path support in the block manager, this commit exposes
multiple data directories in the FsManager and to rest of Kudu. As a bonus,
we can now set an explicit WAL directory and thus isolate log segments to
their own disk.

The FsManager now manages a set of "filesystem roots" consisting of the WAL
path, the data paths, and a metadata path. In tests, it's OK to use one path
for all of these. In production, it's expected that the WAL and data paths
are explicitly specified; the metadata path is simply the first data path.
Each filesystem root has an 'instance' file (with the same UUID throughout).

So what does the filesystem look like?
- Metadata path: instance, master-blocks/, data/ (for superblocks only),
  and consensus-meta/.
- Data paths: instance, data/ (for normal data blocks).
- WAL path: instance, wals/.

The master and TS now have --{master,tablet_server}_wal_dir and
--{master,tablet_server}_data_dirs gflags with which the paths can be set.
All tests squash everything into one path, regardless of whether the test
uses its own FsManager, a MiniCluster, or an ExternalMiniCluster.

Change-Id: Ie56e8ae45f100191a389986e1b31ea9fbb32303e
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5959
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>