Add checksum scanner impl as a TS RPC method
This allows a client to send a request to get a checksum of a tablet.
The checksums are done with a per-row CRC32C that are then added
together across all rows in the tablet.

At this time, the timestamped snapshot approach is untested. This
currently requires READ_LATEST on a quiesced tablet to get a result that
can be compared across tablets.

Change-Id: I8fa0270f8426f13f6e7e1fdcc43c5ad49a3580a8
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5886
Tested-by: Michael Percy <mpercy@cloudera.com>
Reviewed-by: Todd Lipcon <todd@cloudera.com>