Rename TabletServerTest->TabletServerTestBase and make children start the TS themselves
Some tests which inherit from TabletServerTest do not need to start a tablet server,
for multiple reasons. In these cases a tablet server is started which is not used,
logging misleading information.

This makes it so that child classes start the tablet server explicitly if they so
require, this also renames TabletServerTest to TabletServerTestBase since it's no
longer used as a test by itself.

Change-Id: I478cf0c1fa7f5c092987cdb6a2efc1d6a370343b
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5968
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: jenkins