KUDU-452: persist orphaned data blocks in the superblock
This is the "GC" part of KUDU-452. When flushing a superblock, we persist
all blocks now known to be orphaned. Later, when we read the superblock from
disk, we delete those blocks if we can find them.

As part of this change, I've rewritten the orphaned block deletion logic and
centralized it within TabletMetadata. One of the interesting bits is the
orphaned block lifecycle. Orphaned blocks are created when they disappear
from the superblock, and "forgotten" only after they're gone from the
filesystem or have been successfully deleted. Meaning, a block that fails to
delete over and over will never be forgotten; this is a feature not a bug.

There are no crash tests for the persistence logic; that will come later.

Change-Id: I86a2efe9ab092220145c330da670e45ce9d24f5b
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5981
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>