Rename KuduScanner::SetSnapshot()-> SetSnapshotMicros() and add a SetSnapshotRaw() method
When we set a snapshot in a scanner usually we want to pick a physical time for
that snapshot to be executed, which is what SetSnapshot() expects. However sometimes
we might want to pass an actual HybridTime timestamp, e.g. if the scan is caused by
a write and we want to enforce read-your-writes. In this case the scanner should be
able to accept a raw timestamp (one that also includes the logical component).

Change-Id: Icbd4a605949259f10c1a13d83d0a847931e2e684
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5980
Tested-by: jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>