Fix unit test failures when using log block manager
RemoteBootstrapSessionTest failed because the TabletHarness' FsManager was
destroyed before its Tablet. As a result, we lost the block manager and
containers, so that when we tried to destroy the Tablet and its open blocks,
we couldn't punch holes in containers whose blocks had been deleted.

An alternative is for containers to be reference counted, but requiring
clients to destroy all blocks before destroying their owning block manager
doesn't seem like such a bad contract to me.

Separately, I noticed that a SyncDir() call in the log block manager wasn't
respecting the fsync gflag, so I fixed that, and also fixed the bypass of
the SyncMetadata()/SyncContainer() code paths (bypassing them meant
dirty_dirs_ grew unbounded).

Change-Id: I3b947c4c4c1e75c6c9229df51685e7d3bfeee539
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5985
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>