tablet: make bloom filter block size configurable, change default
JD and I noticed in the ITBLL cluster that, after there are a couple billion
rows, the bloom filters no longer fit in the in-process block cache. This was
expected -- 1B rows per server requires ~1.2GB of bloom filter, but the block
cache is configured to only 512MB on that cluster.  After this happened, the
performance became dominated by system CPU in reading bloom filter blocks into
the block cache -- even though they were in the OS cache, the kernel had to
copy them into our own block cache.

After pondering it for a bit, we realized that, for a 64KB bloom block (the
old default), we were copying 64KB of data where we only needed to read
10 bits of it. This is a lot of CPU blow-up for this workload. On the cluster,
we tested using 4KB bloom blocks, and after inserting 1B rows, the throughput
was about 3x better than with the 64KB blocks.

The trade-off here is that we now have a deeper index on the bloom filters, which
means we spend a bit more CPU in index traversal and look up slightly more index
blocks in the cache on average. In the uniform random write workload, these extra
costs are completely outweighed by the benefits described above.

Change-Id: I7213c4f463fa67874d30ed3030e9c481e4a759f3
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5988
Reviewed-by: Jean-Daniel Cryans <jdcryans@cloudera.com>
Tested-by: Todd Lipcon <todd@cloudera.com>