Assertion that compactions don't happen in the past
This adds an assertion that, when we construct a compaction input,
the snapshot we use is newer than any previously-flushed deltas.
This exposes a bug which can cause UNDO deltas to get counted as REDO
deltas and cause missing rows and other bad behavior.

This also hacks linked_list-test to reproduce the issue pretty reliably
in slow test mode.

Change-Id: I845fd1a155220f67a93f3c2ca15df9ff3c8b115b
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5995
Tested-by: jenkins <cloudera-hudson@gerrit.cloudera.org>
Reviewed-by: David Alves <david.alves@cloudera.com>