env: add IsDirectory and Walk functions
I didn't end up needing IsDirectory() but I kept it in since it's got unit
tests and may be useful in the future.

Walk(), however, is interesting and useful. At first I wrote a stack-based
implementation using GetChildren(), then I saw DeleteRecursively(). So I
built Walk() using the guts of DeleteRecursively() and reimplemented the
latter as a client of Walk(). I chose to expose only the bare minimum of fts
concepts to Walk() callers; e.g. no differentiation between files and
symlinks. That can be added as the need arises.

Change-Id: I18f7161862ff9ba47363b3ef6e84aef6b88b8fc4
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5996
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Tested-by: Todd Lipcon <todd@cloudera.com>