KUDU-597 part 1: add a test and assertion which detects mis-sequenced replicate/prepare
This adds a small new utility class TransactionOrderVerifier which validates
each operation just before it is submitted to the Apply pool. The
TransactionDriver code is updated to take a monotonic timestamp during the
Prepare phase, and then the order verifier instance verifies that operations
with higher log indexes were also prepared later than any earlier operation.

This also adds a new integration/stress test which hits follower replicas with
a large number of operations while doing leader elections. This triggers the
bug as described in the JIRA, and that now causes the tablet server to fail
rather than to mis-sequence the operations.

The new unit test as well as a couple of existing ones are marked DISABLED
by this patch, since they now fail consistently. A later patch in this series
will re-enable them when the bug is fixed.

Change-Id: Ia01dd85b9076ee6cc3f2b211222c3873d1ca3fcf
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6015
Reviewed-by: David Alves <david.alves@cloudera.com>
Tested-by: jenkins <cloudera-hudson@gerrit.cloudera.org>