KUDU-597 part 2: Move creation of ConsensusRound into TransactionDriver::Init
This moves some more work into the parallel path instead of the Prepare()
thread, and removes a few lines of code in the process. It also makes
the TransactionDriver code slightly easier to reason about because
the ConsensusRound object is never NULL once the transaction is in-flight.

This small refactor also makes the fix for KUDU-597 slightly cleaner: we can
now use the ConsensusRound to stash consensus-specific information about a
transaction from the time it is initialized, even though it has not yet
replicated.

Change-Id: I74344a577bfa5de3fc35830d9fd97f26324d9535
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6029
Tested-by: jenkins <cloudera-hudson@gerrit.cloudera.org>
Reviewed-by: David Alves <david.alves@cloudera.com>