KUDU-597 part 3: fix potential mis-ordering of PREPARE relative to REPLICATE
This fixes an issue where a node could execute operations out-of-order with
respect to the correct replicated order. This would result in replica divergence
in the case that the operations were non-commutative (eg two operations on the
same row). In particular, we saw on the ITBLL cluster on rare occasion that
an INSERT followed by an UPDATE would yield a "NotFound" error for the UPDATE
on one of the replicas, but not on others.

The typical way this occurred was the following:

1) Replica is acting as FOLLOWER in term 1. Assume everyone has replicated
   through operation 1.10.
2) Replica receives an operation from the client.
  - prior to this patch, it would not check LEADER/FOLLOWER status at this
    point. Instead, it just submitted the operation to the Prepare queue.
3) Replica receives an operation 1.11 from the current leader, and adds it to
   the Prepare queue.
4) Replica is elected leader for term 2.
5) The prepare queue processes the operation enqueued in step 2 above, and calls Replicate()
   which succeeds because it is now the leader. The operation is assigned ID 2.12.
6) The replica now processes 1.11 in the prepare queue.

In this case, operation 1.11 is processed after 2.12, which can cause divergence
in the case that they operate on the same row.

The strategy to fix it is reasonably simple:
- in TransactionDriver::ExecuteAsync(), if the operation is being submitted as
  a LEADER operation, we check the leadership at that point in time. If it is
  not the leader, it is rejected immediately. Otherwise, the leader's current
  term is recorded within the ConsensusRound object.
- in RaftConsensus::Replicate(), we check that we are still the leader, and that
  the term is still the same as was recorded when the operation is submitted.
  This double-check avoids a sort of "ABA" race where in fact the replica has
  stepped down and been re-elected during the time that the operation was in the
  queue.

This patch re-enables several integration tests which fail reliably without the
fix.

Change-Id: Ib51a330c8f561bb87d2ee3230189f543855212a1
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6030
Tested-by: jenkins <cloudera-hudson@gerrit.cloudera.org>
Reviewed-by: David Alves <david.alves@cloudera.com>